refactor(validation-rule-list): move getValidationRulesHint
