+++ apps/demo/src/app/examples/with-styled-components/dictionary/date.component.tsx
@@ -1,9 +1,9 @@
 import { FieldErrors, Path, FieldValues } from 'react-hook-form';
+import { Validations } from '@bedrockstreaming/form-builder';
 import {
   getValidationRulesHints,
-  Validations
-} from '@bedrockstreaming/form-builder';
-import { checkRules } from '@bedrockstreaming/form-validation-rule-list';
+  checkRules
+} from '@bedrockstreaming/form-validation-rule-list';
 
 import { ValidatedTextField } from '@forms/examples/styled-inputs';
 import { BirthdateInput } from '@forms/examples/birthdate';
