+++ apps/docsite/src/form/dictionary/date.component.jsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
-import { getValidationRulesHints } from '@bedrockstreaming/form-builder';
 import {
+  getValidationRulesHints,
   checkRules,
   withValidationRuleList
 } from '@bedrockstreaming/form-validation-rule-list';
