+++ libs/form-builder/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/formBuilder';
 export * from './lib/types';
 export * from './lib/constants';
-export { getValidationRulesHints } from './lib/utils/validation.utils';
