+++ libs/form-builder/src/lib/utils/__tests__/validation.utils.spec.js
@@ -1,5 +1,5 @@
 import { rule } from '../rule.utils';
-import { getFieldRules, getValidationRulesHints } from '../validation.utils';
+import { getFieldRules } from '../validation.utils';
 
 jest.mock('../rule.utils');
 
@@ -50,60 +50,6 @@ describe('getFieldRules', () => {
         JSON.stringify({ func1: extraValidation.func1() })
       );
       expect(rules.validate.func2).toBeUndefined();
-    });
-  });
-
-  describe('getValidationRulesHints', () => {
-    const validation = {
-      foo: {
-        key: 'foo',
-        message: 'foo message'
-      },
-      bar: {
-        key: 'bar',
-        message: 'bar message'
-      },
-      required: {
-        key: 'required',
-        message: 'required message'
-      }
-    };
-
-    const config = {};
-
-    const t = jest.fn().mockImplementation((x) => x);
-
-    rule.mockImplementation((x, y) => ({
-      key: x,
-      check: y
-    }));
-
-    beforeEach(() => {
-      t.mockClear();
-      rule.mockClear();
-    });
-
-    it('should pick the custom rules only', () => {
-      expect(getValidationRulesHints({ t, validation })).toEqual([
-        { check: expect.any(Function), key: validation.foo.message },
-        { check: expect.any(Function), key: validation.bar.message }
-      ]);
-    });
-
-    it('should assign a callback that evaluates to false when key is found in errors types', () => {
-      const result = getValidationRulesHints({
-        t,
-        validation,
-        errors: { types: { foo: true } }
-      });
-      expect(result[0].check()).toBeFalsy();
-      expect(result[1].check()).toBeTruthy();
-    });
-
-    it('should translate the message of each rule', () => {
-      getValidationRulesHints({ t, validation, config });
-      expect(t).toBeCalledWith(validation.foo.message, config);
-      expect(t).toBeCalledWith(validation.bar.message, config);
     });
   });
 });
