+++ libs/form-validation-rule-list/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/rule';
 export * from './lib/utils';
 export * from './lib/constants';
+export * from './lib/getValidationRulesHints';
 export * from './lib/components/dotText.component';
 export * from './lib/components/dotTextList.component';
 export * from './lib/components/validationRuleList.component';
